requirements.txt
matplotlib
numpy
pandas
torch
torchvision
Pillow
tqdm
scikit-learn

opencv-python 
pillow 

pyarrow
fastparquet

huggingface_hub[fsspec]
fsspec[http]
aiohttp

albumentations